+++ nuxeo-drive-client/nxdrive/client/remote_document_client.py
@@ -319,10 +319,7 @@ class RemoteDocumentClient(BaseAutomationClient):
             lock_created = parser.parse(lock_created)
 
         # Permissions
-        context_parameters = doc.get('contextParameters')
-        permissions = (context_parameters.get('permissions')
-                       if context_parameters is not None
-                       else None)
+        permissions = doc.get('contextParameters', {}).get('permissions', None)
 
         # XXX: we need another roundtrip just to fetch the parent uid...
         if parent_uid is None and fetch_parent_uid:
@@ -330,15 +327,16 @@ class RemoteDocumentClient(BaseAutomationClient):
 
         # Normalize using NFC to make the tests more intuitive
         if 'uid:major_version' in props and 'uid:minor_version' in props:
-            version = str(props['uid:major_version']) + "." + str(props['uid:minor_version'])
+            version = str(props['uid:major_version']) + '.' + str(props['uid:minor_version'])
         else:
             version = None
         if name is not None:
             name = unicodedata.normalize('NFC', name)
         return NuxeoDocumentInfo(
             self._base_folder_ref, name, doc['uid'], parent_uid,
-            doc['path'], folderish, last_update, lastContributor,
-            digestAlgorithm, digest, self.repository, doc['type'], version, doc['state'], has_blob, filename,
+            doc['path'], folderish, last_update, last_contributor,
+            digest_algorithm, digest, self.repository, doc['type'],
+            version, doc['state'], has_blob, filename,
             lock_owner, lock_created, permissions)
 
     def _filtered_results(self, entries, fetch_parent_uid=True,
@@ -489,84 +487,23 @@ class RemoteDocumentClient(BaseAutomationClient):
     # Nuxeo Drive specific operations
     #
 
-    def get_repository_names(self):
-        return self.execute("GetRepositories")[u'value']
-
     def get_roots(self):
-        entries = self.execute("NuxeoDrive.GetRoots")[u'entries']
+        entries = self.execute('NuxeoDrive.GetRoots')['entries']
         return self._filtered_results(entries, fetch_parent_uid=False)
 
+    def get_update_info(self):
+        return self.execute('NuxeoDrive.GetClientUpdateInfo')
+
     def register_as_root(self, ref):
-        ref = self._check_ref(ref)
-        self.execute("NuxeoDrive.SetSynchronization", op_input="doc:" + ref,
-                     enable=True)
+        self.execute(
+            'NuxeoDrive.SetSynchronization',
+            op_input='doc:' + self._check_ref(ref),
+            enable=True)
         return True
 
     def unregister_as_root(self, ref):
-        ref = self._check_ref(ref)
-        self.execute("NuxeoDrive.SetSynchronization", op_input="doc:" + ref,
-                     enable=False)
+        self.execute(
+            'NuxeoDrive.SetSynchronization',
+            op_input='doc:' + self._check_ref(ref),
+            enable=False)
         return True
-
-    def make_file_in_user_workspace(self, content, filename):
-        """Stream the given content as a document in the user workspace"""
-        file_path = self.make_tmp_file(content)
-        try:
-            return self.execute_with_blob_streaming("UserWorkspace.CreateDocumentFromBlob", file_path,
-                                                    filename=filename)
-        finally:
-            os.remove(file_path)
-
-    def activate_profile(self, profile):
-        self.execute("NuxeoDrive.SetActiveFactories", profile=profile)
-
-    def deactivate_profile(self, profile):
-        self.execute("NuxeoDrive.SetActiveFactories", profile=profile,
-                     enable=False)
-
-    def get_update_info(self):
-        return self.execute("NuxeoDrive.GetClientUpdateInfo")
-
-    def add_to_locally_edited_collection(self, ref):
-        doc = self.execute("NuxeoDrive.AddToLocallyEditedCollection",
-                           op_input="doc:" + self._check_ref(ref))
-        return doc['uid']
-
-    def get_collection_members(self, ref):
-        docs = self.execute("Collection.GetDocumentsFromCollection",
-                           op_input="doc:" + self._check_ref(ref))
-        return [doc['uid'] for doc in docs['entries']]
-
-    def mass_import(self, target_path, nb_nodes, nb_threads=12):
-        tx_timeout = 3600
-        url = self.server_url + 'site/randomImporter/run?'
-        params = {
-            'targetPath': target_path,
-            'batchSize': 50,
-            'nbThreads': nb_threads,
-            'interactive': 'true',
-            'fileSizeKB': 1,
-            'nbNodes': nb_nodes,
-            'nonUniform': 'true',
-            'transactionTimeout': tx_timeout
-        }
-        for param, value in params.iteritems():
-            url += param + '=' + str(value) + '&'
-        headers = self._get_common_headers()
-        headers.update({'Nuxeo-Transaction-Timeout': tx_timeout})
-        try:
-            log.info('Calling random mass importer on %s with %d threads and %d nodes', target_path,
-                     nb_threads, nb_nodes)
-            self.opener.open(urllib2.Request(url, headers=headers), timeout=tx_timeout)
-        except Exception as e:
-            self._log_details(e)
-            raise e
-
-    def wait_for_async_and_ES_indexing(self):
-        tx_timeout = 3600
-        extra_headers = {'Nuxeo-Transaction-Timeout': tx_timeout}
-        self.execute('Elasticsearch.WaitForIndexing', timeout=tx_timeout, extra_headers=extra_headers,
-                     timeoutSecond=tx_timeout, refresh=True)
-
-    def result_set_query(self, query):
-        return self.execute("Repository.ResultSetQuery", query=query)