+++ nuxeo-drive-client/tests/test_sync_roots.py
@@ -1,13 +1,14 @@
 # coding: utf-8
-from nxdrive.client import RemoteDocumentClient
+from tests.common import RemoteDocumentClientForTests
 from tests.common_unit_test import UnitTestCase
 
 
 class TestSyncRoots(UnitTestCase):
 
     def test_register_sync_root_parent(self):
-        remote = RemoteDocumentClient(self.nuxeo_url, self.user_1, u'nxdrive-test-device-1', self.version,
-                                      password=self.password_1, upload_tmp_dir=self.upload_tmp_dir)
+        remote = RemoteDocumentClientForTests(
+            self.nuxeo_url, self.user_1, u'nxdrive-test-device-1', self.version,
+            password=self.password_1, upload_tmp_dir=self.upload_tmp_dir)
         local = self.local_root_client_1
 
         # First unregister test Workspace